rename login&server
