+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/react": "^11.1.1",
     "@testing-library/user-event": "^12.2.0",
+    "@types/react-router-dom": "^5.1.6",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "firebase": "^8.0.1",
@@ -16,6 +17,7 @@
     "react-naver-login": "^0.1.4",
     "react-naver-maps": "0.0.13",
     "react-router-app": "^0.1.6",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
     "socket.io": "^3.0.1",
     "socket.io-client": "^3.0.1",
